+++ library/Freckle/App/Kafka/Producer.hs
@@ -28,7 +28,6 @@ import qualified Freckle.App.Env as Env
 import Freckle.App.OpenTelemetry
 import Kafka.Producer
 import qualified OpenTelemetry.Trace as Trace
-import UnliftIO.Exception (throwString)
 import Yesod.Core.Lens
 import Yesod.Core.Types (HandlerData)
 
@@ -101,8 +100,10 @@ createKafkaProducerPool addresses KafkaProducerPoolConfig {..} =
         kafkaProducerPoolConfigSize
  where
   mkProducer =
-    either throw pure =<< newProducer (brokersList $ toList addresses)
-  throw err = throwString $ "Failed to open kafka producer: " <> show err
+    either
+      (\err -> throwString ("Failed to open kafka producer: " <> show err))
+      pure
+      =<< newProducer (brokersList $ toList addresses)
 
 produceKeyedOn
   :: ( MonadUnliftIO m