+++ library/Freckle/App/Yesod.hs
@@ -12,7 +12,6 @@ import Freckle.App.Stats (HasStatsClient)
 import qualified Freckle.App.Stats as Stats
 import Network.HTTP.Types (ResponseHeaders, status503)
 import qualified Network.Wai as W
-import UnliftIO.Exception (displayException, handleJust)
 import Yesod.Core.Handler (HandlerFor, sendWaiResponse)
 
 -- | Catch 'SqlError' when queries are canceled due to timeout and respond 503
@@ -28,10 +27,11 @@ respondQueryCanceledHeaders
   => ResponseHeaders
   -> HandlerFor site res
   -> HandlerFor site res
-respondQueryCanceledHeaders headers = handleJust queryCanceled $ \ex -> do
-  logError $ "Query canceled" :# ["exception" .= displayException ex]
-  Stats.increment "query_canceled"
-  sendWaiResponse $ W.responseLBS status503 headers "Query canceled"
+respondQueryCanceledHeaders headers handler =
+  catchJust queryCanceled handler $ \ex -> do
+    logError $ "Query canceled" :# ["exception" .= displayException ex]
+    Stats.increment "query_canceled"
+    sendWaiResponse $ W.responseLBS status503 headers "Query canceled"
 
 queryCanceled :: SqlError -> Maybe SqlError
 queryCanceled ex = ex <$ guard (sqlState ex == "57014")